server/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
basicsr>=1.4.2
realesrgan>=0.3.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
